Implemented Transaction Repository

+++ CustomerRecords/Models/Customer.cs
@@ -7,7 +7,7 @@ namespace CustomerRecords.Api.Models
         public Guid Id { get; set; }
         public string? Name { get; set; }
         public string? Description { get; set; }
-        public int Balance { get; set; } = 0;
+        public Decimal Balance { get; set; } = 0;
         public long? PhoneNumber { get; set; }
         public string? Address { get; set; }
         public List<Transaction>? Transactions { get; set; }
